Added fillMissingDataFields to ensure data quality and icons stay between server and local
