+++ frontend/src/utils/default-data/serviceSearchDefault.json
@@ -1,50 +1,44 @@
-[
-  {
-    "id": "address",
+{
+  "address": {
     "name": "Current Address*",
     "value": "",
     "default": "123 Main St, Anytown, USA",
     "required": true,
     "icon": "MdOutlineLocationOn"
   },
-  {
-    "id": "services_needed",
+  "services_needed": {
     "name": "Services Required*",
     "value": "",
     "default": "None",
     "required": true,
     "icon": "MdOutlineStopCircle"
   },
-  {
-    "id": "country_of_origin",
+  "country_of_origin": {
     "name": "Country of Origin",
     "value": "",
     "default": "Afghanistan",
     "required": false,
     "icon": "MdAirplanemodeActive"
   },
-  {
-    "id": "language",
+  "language": {
     "name": "Language Spoken",
     "value": "",
     "default": "Arabic",
     "required": false,
     "icon": "MdOutlineSpatialAudioOff"
   },
-  {
-    "id": "date_entered",
+  "date_entered": {
     "name": "Date entered US",
     "value": "",
     "default": "2025-01-01",
     "required": false,
     "icon": "MdDateRange"
   },
-  {
-    "id": "family_members",
+  "family_members": {
     "name": "Family Size",
     "value": "",
     "default": "2",
     "required": false,
     "icon": "MdOutlinePersonAddAlt"
   }
-]
+}
